+++ CHANGES.md
@@ -1,8 +1,36 @@
 CHANGES
 =======
 
+- `AuthorizationResponse` class
+    * Added `getResources()` method.
+    * Added `setResources(URI)` method.
+
+- `BackchannelAuthenticationCompleteResponse` class
+    * Added `getResources()` method.
+    * Added `setResources(URI)` method.
+
+- `BackchannelAuthenticationResponse` class
+    * Added `getResources()` method.
+    * Added `setResources(URI)` method.
+
+- `DeviceAuthorizationResponse` class
+    * Added `getResources()` method.
+    * Added `setResources(URI)` method.
+
+- `DeviceVerificationResponse` class
+    * Added `getResources()` method.
+    * Added `setResources(URI)` method.
+
 - `ErrorCode` enum
     * Added `invalid_target`.
+
+- `IntrospectionResponse` class
+    * Added `getResources()` method.
+    * Added `setResources(URI)` method.
+
+- `TokenResponse` class
+    * Added `getResources()` method.
+    * Added `setResources(URI)` method.
 
 
 2.61 (2019-12-03)